fix: preserve screenshot diff result contract
