+++ packages/contracts/src/diff.ts
@@ -1,23 +1,16 @@
-export type ArtifactDescriptor = Record<string, unknown>;
-export type ScreenshotDiffResult = {
-  diffPath?: string;
-  totalPixels: number;
-  differentPixels: number;
-  mismatchPercentage: number;
-  match: boolean;
-  currentOverlayPath?: string;
-  currentOverlayRefCount?: number;
-  regions?: Array<{ currentOverlayMatches?: Array<{ ref?: string }> }>;
-};
 export type SnapshotDiffLine = {
   kind: 'added' | 'removed' | 'unchanged';
   text: string;
+  /**
+   * Plain ref body (`e12`) of the current-tree node behind an added line.
+   * Only populated with `withRefs`; removed and unchanged lines never carry it.
+   */
   ref?: string;
 };
-export type SnapshotDiffSummary = { additions: number; removals: number; unchanged: number };
-
-export type DiffScreenshotCommandResult = ScreenshotDiffResult & {
-  artifacts?: ArtifactDescriptor[];
+export type SnapshotDiffSummary = {
+  additions: number;
+  removals: number;
+  unchanged: number;
 };
 
 export type DiffSnapshotCommandResult = {
