+++ src/commands/interaction/runtime/selector-read-utils.ts
@@ -1,4 +1,4 @@
-import type { RefTarget, SelectorTarget } from '../../../contracts/interaction.ts';
+import type { RefTarget, SelectorTarget } from '@agent-device/contracts/interaction';
 import { AppError } from '@agent-device/kernel/errors';
 import type { SnapshotQualityVerdict } from '../../../snapshot/snapshot-quality.ts';
 import type { FindLocator } from '../../../selectors/find.ts';
